(fix) refactor and fix sidebar scroll issue
